backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
asyncpg==0.29.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
python-dotenv==1.0.1
python-multipart==0.0.9
